initialize networkName and move isOnline logic to a hook.
